Switch from hammer.js to hammerjs
The Hammer.js repo is registered with Bower as both `hammer.js` and
`hammerjs`. Both point to the same Github repo. The Github repo has
a `bower.json` file with the name `hammerjs`, causing it to install
in `bower_components/hammerjs`. This patch fixes ember-gestures to
use the correct bower_component path, as otherwise Ember would not
build because it couldn't find `bower_components/hammer.js`.